Move Resources.Services to Resources.Helpers
Refactor Celbridge.Resources by renaming the Services namespace/folder to Helpers. Files were moved from Services/ to Helpers/ and their namespaces updated, and all using directives across commands, tests, GlobalUsings, and ServiceConfiguration were adjusted to reference Celbridge.Resources.Helpers. This reorganizes helper utilities (e.g. AddResourceHelper, ArchiveHelper, EntityFileHelper, FileEditMatching, FileSystemHelper, PathValidator, ResourceUtils) and updates callers so the project compiles with the new namespace structure.
