+++ Source/Tests/Resources/PathValidatorTests.cs
@@ -1,4 +1,4 @@
-using Celbridge.Resources.Services;
+using Celbridge.Resources.Helpers;
 
 namespace Celbridge.Tests.Resources;
