+++ Source/Workspace/Celbridge.Resources/Commands/FileEditCommand.cs
@@ -1,5 +1,6 @@
 using Celbridge.Commands;
 using Celbridge.Logging;
+using Celbridge.Resources.Helpers;
 using Celbridge.Workspace;
 
 namespace Celbridge.Resources.Commands;
