+++ Source/Workspace/Celbridge.Resources/Commands/FileMultiEditCommand.cs
@@ -1,5 +1,6 @@
 using Celbridge.Commands;
 using Celbridge.Logging;
+using Celbridge.Resources.Helpers;
 using Celbridge.Workspace;
 
 namespace Celbridge.Resources.Commands;
