+++ Source/Workspace/Celbridge.Resources/ServiceConfiguration.cs
@@ -1,4 +1,5 @@
 using Celbridge.Resources.Commands;
+using Celbridge.Resources.Helpers;
 using Celbridge.Resources.Services;
 
 namespace Celbridge.Resources;
