+++ Source/Workspace/Celbridge.Resources/Services/ResourceOperations.cs
@@ -1,4 +1,5 @@
 using Celbridge.Entities;
+using Celbridge.Resources.Helpers;
 
 namespace Celbridge.Resources.Services;
